yolov8-pytorch-cdk/pyproject.toml
[project]
name = "yolov8-pytorch-cdk"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aws-cdk-lib==2.250.0",
    "boto3~=1.24.1",
    "constructs<=11.0.0",
    "jsonpickle~=2.2.0",
    "opencv-python>=4.13.0.92",
]
